Moved net_node from offst-node into offst-bin.
