+++ components/bin/src/lib.rs
@@ -11,7 +11,11 @@
 extern crate log;
 
 mod file_trusted_apps;
+mod node;
+pub mod stnodelib;
+
 pub mod stindexlib;
 pub mod stmgrlib;
-pub mod stnodelib;
 pub mod strelaylib;
+
+pub use crate::node::{net_node, NetNodeError};
