+++ components/node/src/lib.rs
@@ -15,11 +15,11 @@ extern crate log;
 extern crate serde;
 
 // mod adapters;
-mod net_node;
+// mod net_node;
 mod node;
 mod types;
 
-pub use self::net_node::{net_node, NetNodeError, TrustedApps};
-pub use self::node::node;
+// pub use self::net_node::{NetNodeError, TrustedApps};
+pub use self::node::{node, NodeError};
 pub use self::types::{NodeConfig, NodeMutation, NodeState};
 pub use app_server::{ConnPairServer, IncomingAppConnection};
