+++ components/stcompact/src/lib.rs
@@ -18,7 +18,7 @@ mod compact_node;
 
 mod gen;
 mod messages;
-mod server_loop;
+// mod server_loop;
 mod store;
 
-mod stcompactlib;
+// mod stcompactlib;
